version 11

+++ app/src/main/java/com/scr/alertix/Utils/SessionManager.java
@@ -10,6 +10,8 @@ public class SessionManager {
     private static final String KEY_USER_ID = "idUsuario";
 
     private static final String  id_Profile="idProfile";
+
+    private static final String KEY_DIRECCION_ID="idDireccion";
     private static final String KEY_IS_LOGGED_IN = "isLoggedIn";
 
     private SharedPreferences pref;
@@ -18,8 +20,10 @@ public class SessionManager {
 
     public SessionManager(Context context) {
         this.context = context;
-        pref = context.getSharedPreferences(PREF_NAME, Context.MODE_PRIVATE);
-        editor = pref.edit();
+        if (context != null) {
+            pref = context.getSharedPreferences(PREF_NAME, Context.MODE_PRIVATE);
+            editor = pref.edit();
+        }
     }
 
     public SessionManager(PublicacionesAdapter publicacionesAdapter) {
@@ -27,26 +31,49 @@ public class SessionManager {
     }
 
     // Guardar sesión
-    public void saveSession(Long idUsuario,int idProfile) {
-        editor.putLong(KEY_USER_ID, idUsuario);
-        editor.putInt(id_Profile, idProfile);
-        editor.putBoolean(KEY_IS_LOGGED_IN, true);
-        editor.apply(); // Guarda los cambios asíncronamente
+    public void saveSession(Long idUsuario, int idProfile, Long idDireccion) {
+        if (editor != null) {
+            editor.putLong(KEY_USER_ID, idUsuario);
+            editor.putInt(id_Profile, idProfile);
+            editor.putLong(KEY_DIRECCION_ID, idDireccion);
+            editor.putBoolean(KEY_IS_LOGGED_IN, true);
+            editor.apply();
+        }
     }
+
+
 
     // Obtener el ID guardado
     public Long getUserId() {
-        return pref.getLong(KEY_USER_ID, 0);
+        if (pref == null) return 0L;
+        try {
+            return pref.getLong(KEY_USER_ID, 0L);
+        } catch (ClassCastException e) {
+            // Manejar caso donde se guardó como Integer en versiones anteriores
+            return (long) pref.getInt(KEY_USER_ID, 0);
+        }
+    }
+    public Long getDireccionId() {
+        if (pref == null) return 0L;
+        try {
+            return pref.getLong("idDireccion", 0L);
+        } catch (ClassCastException e) {
+            // Si antes era Integer, lo recuperamos así y lo convertimos
+            return (long) pref.getInt("idDireccion", 0);
+        }
     }
 
     // Saber si hay una sesión activa
     public boolean isLoggedIn() {
+        if (pref == null) return false;
         return pref.getBoolean(KEY_IS_LOGGED_IN, false);
     }
 
     // Cerrar sesión
     public void logout() {
-        editor.clear();
-        editor.apply();
+        if (editor != null) {
+            editor.clear();
+            editor.apply();
+        }
     }
 }
